pyproject.toml
[tool.poetry]
name = "python-project"
version = "0.1.0"
description = ""
authors = ["hyeonjoo95 <ekznxh@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.2"
uvicorn = "^0.30.6"
sqlalchemy = "^2.0.32"
pytest = "^8.3.2"
httpx = "^0.27.2"
mysql-connector-python = "^9.0.0"
python-dotenv = "^1.0.1"
pymysql = "^1.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
